requirements.txt
torch>=2.1
torchvision>=0.16
kagglehub>=0.3
PyYAML>=6.0
